chore(core): remove legacy token parser
