+++ .changeset/split-token-parser-into-modules.md
@@ -1,0 +1,6 @@
+---
+'@lapidist/design-lint': patch
+---
+
+refactor token parser into modular pipeline
+
